remove nft account service
