+++ src/test/integration/nfts.e2e-spec.ts
@@ -9,12 +9,10 @@ import { NftOwner } from "src/endpoints/nfts/entities/nft.owner";
 import { NftAccount } from "src/endpoints/nfts/entities/nft.account";
 import { GatewayNft } from "src/endpoints/nfts/entities/gateway.nft";
 import userAccount from "../data/accounts/user.account";
-import { NftAccountService } from "src/endpoints/nfts/nft.account.service";
 import { NftModule } from "src/endpoints/nfts/nft.module";
 
 describe('Nft Service', () => {
   let nftService: NftService;
-  let nftAccountService: NftAccountService;
   let nftCreator: string;
 
   const nftAddress: string = 'erd1qqqqqqqqqqqqqpgqrc4pg2xarca9z34njcxeur622qmfjp8w2jps89fxnl';
@@ -29,7 +27,6 @@ describe('Nft Service', () => {
     }).compile();
 
     nftService = moduleRef.get<NftService>(NftService);
-    nftAccountService = moduleRef.get<NftAccountService>(NftAccountService);
 
     const nfts = await nftService.getNfts({ from: 0, size: 1 }, new NftFilter());
     expect(nfts).toHaveLength(1);
@@ -236,7 +233,7 @@ describe('Nft Service', () => {
     it(`should return a list with nfts`, async () => {
       const nftFilter = new NftFilter();
       nftFilter.identifiers = ['LKFARM-9d1ea8-44b421'];
-      const nfts = await nftAccountService.getAccountNfts(nftAddress, nftFilter);
+      const nfts = await nftService.getAccountNfts(nftAddress, nftFilter);
 
       for (const nft of nfts) {
         expect(nft).toHaveStructure(Object.keys(new GatewayNft()));
@@ -248,7 +245,7 @@ describe('Nft Service', () => {
     it(`should return a list with nfts from internal address`, async () => {
       const nftFilter = new NftFilter();
       nftFilter.identifiers = ['LKFARM-9d1ea8-44b421'];
-      const nfts = await nftAccountService.getNftsForAddressInternal(nftAddress, nftFilter);
+      const nfts = await nftService.getNftsForAddressInternal(nftAddress, nftFilter);
 
       for (const nft of nfts) {
         expect(nft).toHaveStructure(Object.keys(new NftAccount()));
